+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.12.0] - 2019-02-27
+### Changed
+- Start using assembly resolvers from `store-graphql`
+
+### Added
+- Show assembly options removed from item by user
+
 ### Fixed
 
 - Fix tests.